@cloud-fan 2023-01-04
```
SELECT *
FROM bf1
    JOIN bf2
    JOIN bf3
    ON bf1.c1 = bf2.c2
        AND bf3.c3 = bf2.c2
WHERE bf2.a2 = 5
```

Can you show the query plan before and after your PR?